# services/web-api-v2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dynamic API Gateway với auto-discovery từ workers",
        "title": "gRPC Hub Dynamic API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v2",
    "paths": {
        "/capabilities": {
            "get": {
                "description": "Get all service capabilities registered by workers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "List all available capabilities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/invoke/{capability}": {
            "post": {
                "description": "Dynamically invoke any registered service capability",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Invoke a service capability",
                "parameters": [
                    {
                        "type": "string",
                        "example": "hello",
                        "description": "Capability name",
                        "name": "capability",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Get current API gateway status and info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get API status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}